Some minor cleaning up.
@executable_path/../Frameworks now gets automatically added when
building QScintilla and Qwt. Something clearly happened recently since
we used to need to add @executable_path/../Frameworks. Maybe the new
version of Qt is somehow responsible for this or maybe CMake (most
likely)? Whatever the case, we must not add
@executable_path/../Frameworks anymore otherwise the build fails with a
message telling us that the RPATH is already set, etc.