Initial .clang-format file (#2061).
Note: we used to disable llvm-include-order in Clang-Tidy, but we
should have it back. This means that some of our files (e.g.
src/plugins/support/CellMLSupport/src/cellmlfile.cpp) will have to be
reformatted a bit so that headers don’t get wrongly sorted (because of
our, say, inclusion of “cellmlapibegin.h” and “cellmlapiend.h”).