Plugin manager: all our solvers should be loaded when in CLI mode (#2234).
This ensures that we can do whatever we want from Python when in "CLI mode". If we are in GUI mode, we want to allow the user to be able to select whatever solver s/he wants to have. When it comes to Python support, if a solver is not loaded but needed by a Python script, then we will get an error (see commit d51e9669d).