Data store interface: got rid of all our assertions.
Indeed, although those methods are not officially part of our Python wrappers, they could still be called from Python. So, we need to ensure that they don't cause problems (if the original assertion was to be false), which is what we are already doing for methods that are officially part of our Python wrappers.