Simulation Experiment view: fixed a small problem with not being able to reset a simulation that generates NANs (#2299).
We used to compare using qIsNull() and then, at some point, we replaced it with qFuzzyCompare() which ought to be the correct way of doing things except that it doesn't work with NAN (and probably INF) values. At least, not in release mode (yes, everything works fine in debug mode!), so we are now also checking for INF and NAN values.