LLVM+Clang: disable support for zlib (#2459).
With LLVM+Clang 12.0.0, zlib support is broken (?) on Linux and macOS (i.e. undefined references to `compress2`, `uncompress`, `compressBound`, and `crc32`) while it's (still?) all fine on Windows.

Either way, we don't zlib for our use in OpenCOR, so we can safely disable it. (Indeed, can't afford the time to find out what is going on wrong with LLVM+Clang and zlib...)