Tool bar widget: provide a default fixed size for a created wheel.
Without this, to switch from one file (in the Simulation Experiment view) to another file (also in the Simulation Experiment view) which view has not yet been initialised then, in the new view, the wheel will start with a default QwtWheel-based size and then have its size properly set. This means that for a split second, the wheel has the wrong size.