CellML test: properly retrieve the id of a CellML/MathML element (#2581).
This is how things should be done, but... due to a bug in QDomNamedNodeMap::namedItemNS(), it doesn't work. Indeed, no matter what the namespace value we use, QDomNamedNodeMap::namedItemNS() will always return an empty string!? That is, if we don't define the namespace in the element for which we want to retrieve the value of an attribute. Then, even in that case, QDomNamedNodeMap::namedItemNS() will only work if the namespace is defined *before* the attribute (see https://bugreports.qt.io/browse/QTBUG-59932)!