Help window: make sure that non-existent pages result in an error message.
Our previous approach relied on QHelpEngineCore::findFile() to return an invalid URL for non-existent pages, but for some reasons this is not working (anymore?) as expected. So, instead, we retrieve the data associated with the requested page and if there is no data then we know that the page doesn't exist... whether it should have existed or not in the first place.